requirements.txt
fastapi==0.108.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pymysql==1.1.0
pandas==2.2.0
pyreadstat==1.2.6
python-dotenv==1.0.1
streamlit==1.32.0
pytest==7.4.4
pytest-cov==4.1.0
requests==2.31.0
httpx==0.26.0
pre-commit==3.5.0
black==23.3.0
flake8==6.0.0
sqlfluff==2.1.1
jupyter==1.0.0